+++ src/hooks/useGcApi.ts
@@ -1,5 +1,7 @@
 import axios from 'axios';
+import Cookies from 'js-cookie';
 import { useGCContext } from '../contexts';
+import { GRAND_CENTRAL_TOKEN_COOKIE } from '../constants/cookie.ts';
 
 export default function useGcApi() {
   const { gcUrl, onGcApiJwtExpire } = useGCContext();
@@ -7,6 +9,14 @@ export default function useGcApi() {
   const instance = axios.create({
     baseURL: gcUrl,
     withCredentials: true,
+  });
+
+  instance.interceptors.request.use(config => {
+    const token = Cookies.get(GRAND_CENTRAL_TOKEN_COOKIE);
+    if (token) {
+      config.headers.Authorization = `Bearer ${Cookies.get(GRAND_CENTRAL_TOKEN_COOKIE)}`;
+    }
+    return config;
   });
 
   if (onGcApiJwtExpire) {