+++ src/types/route.ts
@@ -3,6 +3,6 @@ import React from 'react';
 export type Route = {
   path: string;
   element: React.JSX.Element;
-  label: React.JSX.Element | string;
-  key: string;
+  label?: React.JSX.Element | string;
+  key?: string;
 };